Made table reader
created a function to take the location org table and make it into the
arkham map dictionary. I think this laptop isn't using python3 as the
ignore wildcard is not working, however the unittests are still
passing and it works when imported into ipython. I might look at using
dot to make a map of arkham for documentations sake.